# schemas/v0/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Talaty Scoring API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/datasets/{dataset_id}": {
      "get": {
        "summary": "Get dataset",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {"200": {"description": "OK"}}
      }
    },
    "/datasets/{dataset_id}/versions": {
      "get": {
        "summary": "List dataset versions",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {"200": {"description": "OK"}}
      },
      "post": {
        "summary": "Create dataset version",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {"200": {"description": "OK"}}
      }
    },
...
    "/feature-sets/{feature_set_id}/versions": {
      "get": {
        "summary": "List feature set versions",
        "parameters": [
          {
            "name": "feature_set_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {"200": {"description": "OK"}}
      },
      "post": {
        "summary": "Create feature set version",
        "parameters": [
          {
            "name": "feature_set_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {"200": {"description": "OK"}}
      }
    },
...
  }
}